bug fixes for killer looting feature
- prevent opening/breaking/auto-looting chests if another player already
has the death chest inventory open
- send message and optional sound effect when denied due to death chest
being accessed by another player
- actions on death chest signs are now performed on attached death chest
(breaking, auto-looting), includes testing for open chest inventory and
denying action if appropriate

+++ SavageDeathChest/src/main/java/com/winterhaven_mc/deathchest/BlockEventListener.java
@@ -6,6 +6,7 @@ import java.util.ArrayList;
 
 import org.bukkit.GameMode;
 import org.bukkit.Material;
+import org.bukkit.Sound;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
@@ -69,10 +70,6 @@ public class BlockEventListener implements Listener {
 			// if block is not owned by player, test for override permission or killer-looting enabled
 			if (!block.getMetadata("deathchest-owner").get(0).asString().equals(player.getUniqueId().toString())) {
 				
-				if (plugin.debug) {
-					plugin.getLogger().info("A non-owner is attempting to break a death chest block...");
-				}
-
 				// if player does not have deathchest.loot.other permission,
 				if (!player.hasPermission("deathchest.loot.other")) {
 
@@ -86,9 +83,6 @@ public class BlockEventListener implements Listener {
 							plugin.messageManager.sendPlayerMessage(player,"not-owner");
 							return;
 						}
-						else if (plugin.debug) {
-							plugin.getLogger().info("Death chest breakage allowed by killer-looting configuration setting.");
-						}
 					}
 					else {
 						// killer-looting is not enabled, so send not-owner player message and return
@@ -96,14 +90,39 @@ public class BlockEventListener implements Listener {
 						return;						
 					}
 				}
-				else if (plugin.debug) {
-						plugin.getLogger().info("Death chest breakage allowed by deathchest.loot.other permission.");
-				}
 			}
 		}
+				
+		// if block is wall sign, set block to attached block
+		if (block.getType().equals(Material.WALL_SIGN)) {
+		    Sign sign = (Sign)block.getState().getData();
+		    block = block.getRelative(sign.getAttachedFace());
+		}
+		// if block is sign post, set block to one block below
+		else if (block.getType().equals(Material.SIGN_POST)) {
+			block = block.getRelative(0, 1, 0);
+		}
 		
-		// destroy DeathChestBlock
-		plugin.chestManager.destroyDeathChestBlock(block);
+		// confirm block is a death chest 
+		if (block.getType().equals(Material.CHEST) && DeathChestBlock.isDeathChestBlock(block)) {
+			
+			// if chest inventory has any viewers, output message and return
+			// this is to prevent breaking death chests that are already open by another player
+			if (plugin.chestManager.getChestViewerCount(block) > 0) {
+
+				// send player message
+				plugin.messageManager.sendPlayerMessage(player, "chest-currently-open");
+
+				// play denied access sound
+				if (plugin.getConfig().getBoolean("sound-effects",true)) {
+					player.playSound(player.getLocation(), Sound.VILLAGER_NO, 1, 1);
+				}
+			}
+			else {
+				// destroy DeathChestBlock
+				plugin.chestManager.destroyDeathChestBlock(block);
+			}
+		}
 	}
 
 	
@@ -122,11 +141,6 @@ public class BlockEventListener implements Listener {
 		if (!DeathChestBlock.isDeathChestBlock(block)) {
 			return;
 		}
-		else {
-			if (plugin.debug) {
-				plugin.getLogger().info("BlockDamageEvent detected on death chest block.");
-			}
-		}
 		
 		// if quick-loot is not enabled in configuration, do nothing and return
 		if (!plugin.getConfig().getBoolean("quick-loot", true)) {
@@ -141,7 +155,7 @@ public class BlockEventListener implements Listener {
 		// cancel event
 		event.setCancelled(true);
 		
-		// if player is in creative mode and does not have override permission, cancel event, send message and return
+		// if player is in creative mode and does not have override permission, send message and return
 		if (player.getGameMode().equals(GameMode.CREATIVE) && !player.hasPermission("deathchest.creative-access")) {
 			plugin.messageManager.sendPlayerMessage(player, "no-creative-access");
 			return;
@@ -150,13 +164,10 @@ public class BlockEventListener implements Listener {
 		// if chest-protection is enabled in config, test for ownership
 		if (plugin.getConfig().getBoolean("chest-protection",true)) {
 			
-			// if player is not block owner, test for override permisssion or killer-looting enabled
+			// if player is not block owner, test for override permission or killer-looting enabled
 			if (!block.getMetadata("deathchest-owner").get(0).asString().equals(player.getUniqueId().toString())) {
-				if (plugin.debug) {
-					plugin.getLogger().info("Sneak-puncher is not chest owner.");
-				}
 	
-				// if player does not have deathchest.loot.other permission
+				// if player does not have deathchest.loot.other permission, check if killer looting is enabled
 				if (!player.hasPermission("deathchest.loot.other")) {
 	
 					// if killer-looting is enabled check if player is killer
@@ -168,35 +179,48 @@ public class BlockEventListener implements Listener {
 					
 							// player is not killer, so output message and return
 							plugin.messageManager.sendPlayerMessage(player, "not-owner");
-							if (plugin.debug) {
-								plugin.getLogger().info("Sneak-puncher is not chest owner killer.");
-							}
 							return;
 						}
 					}
 					else {
 						// killer looting is not enabled, so output message and return
 						plugin.messageManager.sendPlayerMessage(player, "not-owner");
-						if (plugin.debug) {
-							plugin.getLogger().info("Killer looting is not enabled in config.");
-						}
 						return;						
 					}
-				}
-				else if (plugin.debug) {
-					plugin.getLogger().info("Sneak-puncher has deathchest.loot.other permission.");
 				}
 			}
 		}
 		
-		// if player clicked sign, set block to attached block
-		if (block.getType().equals(Material.WALL_SIGN) || block.getType().equals(Material.SIGN_POST)) {
+		// if player clicked attached wall sign, set block to attached block
+		if (block.getType().equals(Material.WALL_SIGN)) {
 			Sign sign = (Sign)block.getState().getData();
 			block = block.getRelative(sign.getAttachedFace());
 		}
+		// if player clicked sign post, set block to block below sign post
+		else if (block.getType().equals(Material.SIGN_POST)) {
+			block = block.getRelative(0, -1, 0);
+		}
 		
-		// loot chest
-		plugin.chestManager.lootChest(player, block);
+		// confirm block is a chest
+		if (block.getType().equals(Material.CHEST)) {
+			
+			// if chest inventory has any viewers, output message and return
+			// this is to prevent quick-looting death chests that are already open by another player
+			if (plugin.chestManager.getChestViewerCount(block) > 0) {
+
+				// send player message
+				plugin.messageManager.sendPlayerMessage(player, "chest-currently-open");
+
+				// play denied access sound
+				if (plugin.getConfig().getBoolean("sound-effects",true)) {
+					player.playSound(player.getLocation(), Sound.VILLAGER_NO, 1, 1);
+				}
+			}
+			else {
+				// loot chest
+				plugin.chestManager.lootChest(player, block);
+			}
+		}
 	}
 
 
@@ -262,4 +286,5 @@ public class BlockEventListener implements Listener {
 		// destroy DeathChestBlock
 		plugin.chestManager.destroyDeathChestBlock(block);
 	}
+	
 }
